Type safety interface implement
